Move fix-fsselection parts to libraries
