+++ bin/gftools-fix-fsselection.py
@@ -21,6 +21,9 @@ import sys
 
 from fontTools import ttLib
 import tabulate
+from gftools.utils import get_fsSelection_byte1, get_fsSelection_byte2
+from gftools.util.styles import STYLE_NAMES, is_filename_canonical
+from gftools.fix import fix_fsselection
 
 parser = argparse.ArgumentParser(description='Print out fsSelection'
                                              ' bitmask of the fonts')
@@ -28,38 +31,6 @@ parser.add_argument('font', nargs="+")
 parser.add_argument('--csv', default=False, action='store_true')
 parser.add_argument('--usetypometrics', default=False, action='store_true')
 parser.add_argument('--autofix', default=False, action='store_true')
-
-STYLE_NAMES = ["Thin",
-               "ExtraLight",
-               "Light",
-               "Regular",
-               "Medium",
-               "SemiBold",
-               "Bold",
-               "ExtraBold",
-               "Black",
-               "Thin Italic",
-               "ExtraLight Italic",
-               "Light Italic",
-               "Italic",
-               "Medium Italic",
-               "SemiBold Italic",
-               "Bold Italic",
-               "ExtraBold Italic",
-               "Black Italic"]
-
-RIBBI_STYLE_NAMES = ["Regular",
-                     "Italic",
-                     "Bold",
-                     "BoldItalic"]
-
-
-def getByte2(ttfont):
-  return ttfont['OS/2'].fsSelection >> 8
-
-
-def getByte1(ttfont):
-  return ttfont['OS/2'].fsSelection & 255
 
 
 def printInfo(fonts, print_csv=False):
@@ -70,8 +41,8 @@ def printInfo(fonts, print_csv=False):
     row = [os.path.basename(font)]
     row.append(('{:#010b} '
                 '{:#010b}'
-                '').format(getByte2(ttfont),
-                           getByte1(ttfont)).replace('0b', ''))
+                '').format(get_fsSelection_byte2(ttfont),
+                           get_fsSelection_byte1(ttfont)).replace('0b', ''))
     rows.append(row)
 
   def as_csv(rows):
@@ -85,40 +56,6 @@ def printInfo(fonts, print_csv=False):
   else:
     print(tabulate.tabulate(rows, headers, tablefmt="pipe"))
 
-def get_stylename(filename):
-  filename_base = filename.split('.')[0]
-  return filename_base.split('-')[-1]
-
-def _familyname(filename):
-  filename_base = filename.split('.')[0]
-  names = filename_base.split('-')
-  names.pop()
-  return '-'.join(names)
-
-def is_italic(stylename):
-  return 'Italic' in stylename
-
-def is_regular(stylename):
-  return ("Regular" in stylename or
-          (stylename in STYLE_NAMES and
-           stylename not in RIBBI_STYLE_NAMES and
-           "Italic" not in stylename))
-
-def is_bold(stylename):
-  return stylename in ["Bold", "BoldItalic"]
-
-def is_canonical(filename):
-  if '-' not in filename:
-    return False
-  else:
-    style = get_stylename(filename)
-    for valid in STYLE_NAMES:
-      valid = ''.join(valid.split(' '))
-      if style == valid:
-        return True
-    # otherwise:
-    return False
-
 def main():
   args = parser.parse_args()
   if args.autofix:
@@ -127,38 +64,13 @@ def main():
       ttfont = ttLib.TTFont(font)
       filename = os.path.basename(font)
 
-      if not is_canonical(filename):
+      if not is_filename_canonical(filename):
         print(f"Font filename '{filename}' is not canonical!\n\n"
               f"Filename must be structured as familyname-style.ttf and "
               f"the style must be any of the following {STYLE_NAMES}")
         exit(-1)
 
-      stylename = get_stylename(filename)
-
-      initial_value = ttfont['OS/2'].fsSelection
-
-      if is_regular(stylename):
-        ttfont['OS/2'].fsSelection |= 0b1000000
-      else:
-        ttfont['OS/2'].fsSelection &= ~0b1000000
-
-      if is_bold(stylename):
-        ttfont['OS/2'].fsSelection |= 0b100000
-      else:
-        ttfont['OS/2'].fsSelection &= ~0b100000
-
-      if is_italic(stylename):
-        ttfont['OS/2'].fsSelection |= 0b1
-      else:
-        ttfont['OS/2'].fsSelection &= ~0b1
-
-      if args.usetypometrics:
-        ttfont['OS/2'].version = 4
-        ttfont['OS/2'].fsSelection |= 0b10000000
-      else:
-        ttfont['OS/2'].fsSelection &= ~0b10000000
-
-      if ttfont['OS/2'].fsSelection != initial_value:
+      if fix_fsselection(ttfont, filename):
         fixed_fonts.append(font)
         ttfont.save(font + '.fix')
